+++ __tests__/respect/x-security-bearer-auth/x-security-bearer-auth.arazzo.yaml
@@ -6,7 +6,7 @@ info:
 sourceDescriptions:
   - name: museum-api
     type: openapi
-    url: ../../../resources/museum.yaml
+    url: ../museum-api.yaml
 
 workflows:
   - workflowId: bearer-auth-workflow
