+++ __tests__/respect/x-security-oauth2-auth/x-security-oauth2-auth.arazzo.yaml
@@ -6,7 +6,7 @@ info:
 sourceDescriptions:
   - name: museum-api
     type: openapi
-    url: ../../../resources/museum.yaml
+    url: ../museum-api.yaml
 
 workflows:
   - workflowId: oauth2-auth-workflow
