+++ __tests__/respect/x-security-open-id-connect-auth/x-security-open-id-connect-auth.arazzo.yaml
@@ -6,7 +6,7 @@ info:
 sourceDescriptions:
   - name: museum-api
     type: openapi
-    url: ../../../resources/museum.yaml
+    url: ../museum-api.yaml
 
 workflows:
   - workflowId: open-id-connect-auth-workflow
