+++ __tests__/respect/x-security-scheme-apply-order/x-security-scheme-apply-order.arazzo.yaml
@@ -6,7 +6,7 @@ info:
 sourceDescriptions:
   - name: museum-api
     type: openapi
-    url: ../../../resources/museum.yaml
+    url: ../museum-api.yaml
 
 workflows:
   - workflowId: events-crud
