+++ packages/respect-core/src/modules/__tests__/arazzo-description-generator/generate-arazzo-description.test.ts
@@ -1,5 +1,5 @@
 import { generateArazzoDescription } from '../../arazzo-description-generator/index.js';
-import { type ParameterWithIn } from '../../config-parser/index.js';
+import { type ParameterWithIn } from '../../context-parser/index.js';
 import {
   bundleOpenApi,
   getOperationFromDescription,
