+++ packages/respect-core/src/modules/__tests__/config-parser/handle-request-body-replacements.test.ts
@@ -1,5 +1,5 @@
 import type { RuntimeExpressionContext } from '../../../types.js';
-import { handlePayloadReplacements } from '../../config-parser/index.js';
+import { handlePayloadReplacements } from '../../context-parser/index.js';
 
 describe('handlePayloadReplacements', () => {
   it('should replace the value using JSON Pointer', () => {
