+++ packages/respect-core/src/modules/context-parser/resolve-reusable-object-reference.ts
