+++ packages/respect-core/src/modules/description-parser/get-request-data-from-openapi.ts
@@ -4,10 +4,10 @@ import {
   generateExampleValue,
 } from '../arazzo-description-generator/index.js';
 import { extractFirstExample } from './extract-first-example.js';
-import { isParameterWithIn } from '../config-parser/index.js';
+import { isParameterWithIn } from '../context-parser/index.js';
 
 import type { Parameter } from '../../types.js';
-import type { ParameterWithIn } from '../config-parser/index.js';
+import type { ParameterWithIn } from '../context-parser/index.js';
 import type { OperationDetails } from './get-operation-from-description.js';
 
 export interface OpenApiRequestData {
