+++ packages/respect-core/src/modules/flow-runner/prepare-request.ts
@@ -8,14 +8,14 @@ import {
   resolveReusableComponentItem,
   isParameterWithIn,
   handlePayloadReplacements,
-} from '../config-parser/index.js';
+} from '../context-parser/index.js';
 import { getServerUrl } from './get-server-url.js';
 import { createRuntimeExpressionCtx, collectSecretFields } from './context/index.js';
 import { evaluateRuntimeExpressionPayload } from '../runtime-expressions/index.js';
 import { resolveXSecurityParameters } from './resolve-x-security-parameters.js';
 
 import type { Oas3SecurityScheme } from 'core/src/typings/openapi.js';
-import type { ParameterWithIn } from '../config-parser/index.js';
+import type { ParameterWithIn } from '../context-parser/index.js';
 import type { TestContext, Step, Parameter, PublicStep } from '../../types.js';
 import type { OperationDetails } from '../description-parser/index.js';
